serve ftp: fix race condition when using the auth proxy
In this commit we introduced a race condition when using the auth
proxy.

6a5e1d7bb37e36e3 serve ftp: update to goftp.io/server v2.0.1

This was due to the re-organisation of the upstream library which made
the driver be a singleton rather than per session.

This means that when using the auth proxy we need to keep track of
which VFS to use by based on which FTP user is connected.

This also adjusts the locking so that the methods will run
concurrently.